move configuration constants to env

+++ ui/docker/stop.sh
@@ -1,2 +1,10 @@
 #!/bin/bash
-docker-compose down
+
+if [ -f .env ];
+then
+    set -a
+    source .env
+    docker-compose down
+else
+    echo "Please copy .env.sample to .env"
+fi
